+++ musica/mechanism_configuration/__init__.py
@@ -1,0 +1,1 @@
+from .mechanism_configuration import *
