+++ musica/mechanism_configuration/mechanism_configuration_condensed_phase_arrhenius.py
@@ -1,41 +1,30 @@
 from typing import Optional, Any, Dict, List, Union, Tuple
-from musica import _Arrhenius, _ReactionComponent
-from musica.mechanism_configuration_phase import Phase
-from musica.mechanism_configuration_species import Species
-from musica.mechanism_configuration_reactions import ReactionComponentSerializer
-from musica.mechanism_configuration_utils import add_other_properties, remove_empty_keys
+from musica import _CondensedPhaseArrhenius, _ReactionComponent
+from .mechanism_configuration_phase import Phase
+from .mechanism_configuration_species import Species
+from .mechanism_configuration_reactions import ReactionComponentSerializer
+from .mechanism_configuration_utils import add_other_properties, remove_empty_keys
 
 BOLTZMANN_CONSTANT_J_K = 1.380649e-23  # J K-1
 
 
-class Arrhenius(_Arrhenius):
+class CondensedPhaseArrhenius(_CondensedPhaseArrhenius):
     """
-    A class representing an Arrhenius rate constant.
-
-    k = A * exp( C / T ) * ( T / D )^B * exp( 1 - E * P )
-
-    where:
-        k = rate constant
-        A = pre-exponential factor [(mol m-3)^(n-1)s-1]
-        B = temperature exponent [unitless]
-        C = exponential term [K-1]
-        D = reference temperature [K]
-        E = pressure scaling term [Pa-1]
-        T = temperature [K]
-        P = pressure [Pa]
-        n = number of reactants
+    A class representing a condensed phase Arrhenius rate constant.
 
     Attributes:
-        name (str): The name of the Arrhenius rate constant.
-        A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1] where n is the number of reactants.
+        name (str): The name of the condensed phase Arrhenius rate constant.
+        A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
         B (float): Temperature exponent [unitless].
         C (float): Exponential term [K-1].
+        Ea (float): Activation energy [J molecule-1].
         D (float): Reference Temperature [K].
         E (float): Pressure scaling term [Pa-1].
         reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
         products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the Arrhenius rate constant.
+        aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
+        aerosol_phase_water (Species): The water species in the aerosol phase.
+        other_properties (Dict[str, Any]): A dictionary of other properties of the condensed phase Arrhenius rate constant.
     """
 
     def __init__(
@@ -49,24 +38,26 @@ class Arrhenius(_Arrhenius):
         E: Optional[float] = None,
         reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
         products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
+        aerosol_phase: Optional[Phase] = None,
+        aerosol_phase_water: Optional[Species] = None,
         other_properties: Optional[Dict[str, Any]] = None,
     ):
         """
-        Initializes the Arrhenius object with the given parameters.
+        Initializes the CondensedPhaseArrhenius object with the given parameters.
 
         Args:
-            name (str): The name of the Arrhenius rate constant.
-            A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1] where n is the number of reactants.
+            name (str): The name of the condensed phase Arrhenius rate constant.
+            A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
             B (float): Temperature exponent [unitless].
             C (float): Exponential term [K-1].
             Ea (float): Activation energy [J molecule-1].
             D (float): Reference Temperature [K].
             E (float): Pressure scaling term [Pa-1].
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the Arrhenius rate constant.
+            reactants (List[Union[Species, Tuple[float, Species]]]]: A list of reactants involved in the reaction.
+            products (List[Union[Species, Tuple[float, Species]]]]: A list of products formed in the reaction.
+            aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
+            aerosol_phase_water (Species): The water species in the aerosol phase.
+            other_properties (Dict[str, Any]): A dictionary of other properties of the condensed phase Arrhenius rate constant.
         """
         super().__init__()
         self.name = name if name is not None else self.name
@@ -101,13 +92,16 @@ class Arrhenius(_Arrhenius):
             if products is not None
             else self.products
         )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
+        self.aerosol_phase = aerosol_phase.name if aerosol_phase is not None else self.aerosol_phase
+        self.aerosol_phase_water = (
+            aerosol_phase_water.name if aerosol_phase_water is not None else self.aerosol_phase_water
+        )
         self.other_properties = other_properties if other_properties is not None else self.other_properties
 
     @staticmethod
     def serialize(cls) -> Dict:
         serialize_dict = {
-            "type": "ARRHENIUS",
+            "type": "CONDENSED_PHASE_ARRHENIUS",
             "name": cls.name,
             "A": cls.A,
             "B": cls.B,
@@ -115,8 +109,9 @@ class Arrhenius(_Arrhenius):
             "D": cls.D,
             "E": cls.E,
             "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "gas phase": cls.gas_phase,
+            "products":  ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
+            "aerosol phase": cls.aerosol_phase,
+            "aerosol-phase water": cls.aerosol_phase_water,
         }
         add_other_properties(serialize_dict, cls.other_properties)
         return remove_empty_keys(serialize_dict)
